# [Core] Refactor GGUF parameters packing and forwarding