--- a/vllm/model_executor/layers/linear.py
+++ b/vllm/model_executor/layers/linear.py
@@ -455,7 +455,7 @@ class MergedColumnParallelLinear(ColumnParallelLinear):
             param.shard_id_map[loaded_shard_id] = len(param.data_container)
             param.data_container.append(loaded_weight)
             if len(param.data_container) == 2:
-                param = param.materialize_nested()
+                self.qweight = param.materialize_nested()
             return
 
         param_data = param.data