@Isotr0py 2024-09-28
Seems that the performance has minor improvement after these cleanup:

**Benchmark with Meta-Llama-3.1-8B-Instruct-IQ4_XS on T4**
```shell
python benchmarks/benchmark_throughput.py --input-len 16 --output-len 512 --max-model-len 4096 --num-prompts 100 \
                                          --model ./Meta-Llama-3.1-8B-Instruct-IQ4_XS.gguf \
                                          --tokenizer unsloth/Meta-Llama-3.1-8B-Instruct 
```

**Results**
| Metrics    | Before          | After           |   |
|------------|-----------------|-----------------|---|
| VRAM Usage | 4.2861 GB       | 4.2074 GB       | ↓ |
| Throughout | 261.67 tokens/s | 266.54 tokens/s | ↑ |