@kylesayrs 2024-10-01
Initial evaluation regression looks good

main
```
vllm (pretrained=/home/ksayers/.cache/huggingface/hub/models--TheBloke--TinyLlama-1.1B-Chat-v1.0-GGUF/snapshots/52e7645ba7c309695bec7ac98f4f005b139cf465/tinyllama-1.1b-chat-v1.0.Q4_0.gguf,tokenizer=TinyLlama/TinyLlama-1.1B-Chat-v1.0,dtype=auto,add_bos_token=True,gpu_memory_utilization=0.8,tensor_parallel_size=1), gen_kwargs: (None), limit: None, num_fewshot: 5, batch_size: auto
|  Tasks   |Version|Filter|n-shot|Metric|   |Value |   |Stderr|
|----------|------:|------|-----:|------|---|-----:|---|-----:|
|winogrande|      1|none  |     5|acc   |↑  |0.6133|±  |0.0137|
```

Isotr0py:gguf-optimize
```
vllm (pretrained=/home/ksayers/.cache/huggingface/hub/models--TheBloke--TinyLlama-1.1B-Chat-v1.0-GGUF/snapshots/52e7645ba7c309695bec7ac98f4f005b139cf465/tinyllama-1.1b-chat-v1.0.Q4_0.gguf,tokenizer=TinyLlama/TinyLlama-1.1B-Chat-v1.0,dtype=auto,add_bos_token=True,gpu_memory_utilization=0.8,tensor_parallel_size=1), gen_kwargs: (None), limit: None, num_fewshot: 5, batch_size: auto
|  Tasks   |Version|Filter|n-shot|Metric|   |Value |   |Stderr|
|----------|------:|------|-----:|------|---|-----:|---|-----:|
|winogrande|      1|none  |     5|acc   |?  |0.6133|?  |0.0137|
```